Improve `retention-days` documentation
Clarify that `retention-days: 0` is a valid configuration which is numerically less than 1, but that the shortest retention is `1` (it is _not_ a minimum value according to Math/Numbers/Integers, it's just the shortest retention value).
Clarify that the maximum `retention-days` value is configured per repository (which implicitly inherits from organizations...) but that the default for repositories is 90 days.